requirements.txt
Flask
PyMuPDF
pytesseract
opencv-python
pdf2image
numpy
Pillow
firebase-admin
firebase-functions
requests
Flask
PyMuPDF
pytesseract
opencv-python
numpy
pdf2image
Pillow
fpdf
